backend/felicity_lims/pyproject.toml
[tool.poetry]
name = "FelicityLims"
version = "0.1.0"
description = "Laboratory Information Management System"
authors = ["Aurthur Musendame <aurthurmusendame@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
psycopg2-binary = "^2.9.1"
python-multipart = "^0.0.5"
emails = "^0.6"
pydantic = {extras = ["email"], version = "^1.8.2"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
alembic = "^1.7.3"
graphene-sqlalchemy = "^2.3.0"
sqlalchemy_mixins = "^1.5"
sqlalchemy_mptt ="^0.2.5"
tenacity = "^8.0.1"
apscheduler = "^3.7.0"
pytest = "^5.4.1"
asyncpg = "^0.24.0"
sqlmodel = "^0.0.4"

[tool.poetry.dev-dependencies]
mypy = "^0.770"
black = "^19.10b0"
isort = "^4.3.21"
autoflake = "^1.3.1"
flake8 = "^3.7.9"
pytest = "^5.4.1"
sqlalchemy-stubs = "^0.3"
pytest-cov = "^2.8.1"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"